refactor: dedupe android talk config parsing
(cherry picked from commit cee2f3e8b4c0e7d4eb69f554fa51ea635ed9c49f)
